Rahmat will be updating this to the Final PDF final
I will be adding section 5.1.6 and additional commits regarding the
conclusion tonight and the correct reference format. Cole please do not
update this to the final pdf file because Rahmat will be doing it. Also,
please inform me when everything is update for the pdf file as I will be
doing the final edit, thanks!